Changed builtins to use new command style.
As a piecemeal change to make plugins use new command style started with builtins. Now commands are saved in the event object by the plugin. When a command is received, the string following the ! is checked for existance in each plugin, then all of the rest is split on the space and passed to the command as an array of arguments.